requirements.txt
celery==5.6.3
dj_database_url==3.1.2
Django==5.2
Requests==2.33.0
Scrapy==2.14.2
urllib3==2.6.3
django-debug-toolbar==6.2.0
django-celery-results==2.6.0
django-celery-beat==2.9.0
whitenoise==6.12.0
redis==7.4.0
python-decouple==3.8
gunicorn==23.0.0
